requirements.txt
beautifulsoup4==4.10.0
requests==2.26.0
psycopg2-binary==2.9.3
validators==0.18.2
